refactor(models): Move all Pydantic models to models/ directories
Move 31 Pydantic models to comply with ONEX directory structure standards:

**PR #18 Models (3 models)**:
- ModelEmbeddingHttpClientConfig → models/config/
- ModelRateLimiterConfig → models/config/
- EmbeddingProviderType → EnumEmbeddingProviderType (ONEX naming)

**Utils Models (18 models)**:
- Retry, concurrency, resource manager, observability, audit, health, PII models

**Adapter Models (6 models)**:
- Graph memory models moved to models/adapters/

**Node Models (7 models)**:
- Moved to respective nodes/*/models/ directories

Add pre-commit validation hook (validate_model_locations.py).

Note: Some backward compat aliases remain - to be removed in follow-up per
CLAUDE.md NO BACKWARDS COMPATIBILITY policy.
