+++ src/omnimemory/models/config/__init__.py
@@ -8,21 +8,36 @@ Backend Configs:
 - ModelFilesystemConfig: Local filesystem storage (required for Phase 1)
 - ModelPostgresConfig: PostgreSQL persistent storage (optional)
 - ModelQdrantConfig: Qdrant vector storage (optional)
+- ModelEmbeddingHttpClientConfig: Embedding HTTP client (optional)
 
 Service Config:
 - ModelMemoryServiceConfig: Top-level config composing all backends
 """
 
+from omnimemory.models.config.model_embedding_config import (
+    EnumEmbeddingProviderType,
+    ModelEmbeddingHttpClientConfig,
+)
 from omnimemory.models.config.model_filesystem_config import ModelFilesystemConfig
 from omnimemory.models.config.model_memory_service_config import (
     ModelMemoryServiceConfig,
 )
 from omnimemory.models.config.model_postgres_config import ModelPostgresConfig
 from omnimemory.models.config.model_qdrant_config import ModelQdrantConfig
+from omnimemory.models.config.model_rate_limiter_config import (
+    DEFAULT_REQUESTS_PER_MINUTE,
+    DEFAULT_TOKENS_PER_MINUTE,
+    ModelRateLimiterConfig,
+)
 
 __all__ = [
+    "DEFAULT_REQUESTS_PER_MINUTE",
+    "DEFAULT_TOKENS_PER_MINUTE",
+    "EnumEmbeddingProviderType",
+    "ModelEmbeddingHttpClientConfig",
     "ModelFilesystemConfig",
     "ModelMemoryServiceConfig",
     "ModelPostgresConfig",
     "ModelQdrantConfig",
+    "ModelRateLimiterConfig",
 ]
