+++ src/omnimemory/models/utils/model_concurrency.py
@@ -1,0 +1,43 @@
+"""
+Concurrency-related Pydantic models for OmniMemory ONEX architecture.
+
+This module contains models for connection pool configuration.
+"""
+
+from __future__ import annotations
+
+from pydantic import BaseModel, Field
+
+__all__ = [
+    "ModelConnectionPoolConfig",
+    # Backward compatibility alias
+    "ConnectionPoolConfig",
+]
+
+
+class ModelConnectionPoolConfig(BaseModel):
+    """Configuration for connection pools."""
+
+    name: str = Field(description="Pool name")
+    min_connections: int = Field(default=1, ge=0, description="Minimum connections")
+    max_connections: int = Field(
+        default=50,
+        ge=1,
+        description="Maximum connections (increased for production load)",
+    )
+    connection_timeout: float = Field(
+        default=30.0, gt=0, description="Connection timeout"
+    )
+    idle_timeout: float = Field(
+        default=300.0, gt=0, description="Idle connection timeout"
+    )
+    health_check_interval: float = Field(
+        default=60.0, gt=0, description="Health check interval"
+    )
+    retry_attempts: int = Field(
+        default=3, ge=0, description="Retry attempts for failed connections"
+    )
+
+
+# Backward compatibility alias
+ConnectionPoolConfig = ModelConnectionPoolConfig
