+++ src/omnimemory/nodes/memory_retrieval_effect/handlers/handler_qdrant_mock.py
@@ -66,7 +66,6 @@ from typing import TYPE_CHECKING
 from omnibase_core.models.omnimemory import (
     ModelMemorySnapshot,  # noqa: TC002 - Pydantic needs runtime access
 )
-from pydantic import BaseModel, ConfigDict, Field
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
@@ -75,9 +74,10 @@ from ..clients.embedding_client import (
     EmbeddingClient,
     EmbeddingClientError,
     EmbeddingConnectionError,
-    ModelEmbeddingClientConfig,
 )
 from ..models import (
+    ModelEmbeddingClientConfig,
+    ModelHandlerQdrantMockConfig,
     ModelMemoryRetrievalRequest,
     ModelMemoryRetrievalResponse,
     ModelSearchResult,
@@ -91,58 +91,6 @@ __all__ = [
     "HandlerQdrantMock",
     "ModelHandlerQdrantMockConfig",
 ]
-
-
-class ModelHandlerQdrantMockConfig(BaseModel):
-    """Configuration for the mock Qdrant handler.
-
-    Attributes:
-        embedding_dimension: Dimension of embedding vectors. Defaults to 1024
-            (MLX Qwen3-Embedding compatible).
-        simulate_latency_ms: Simulated latency for operations in milliseconds.
-            Set to 0 for instant responses.
-        use_real_embeddings: Whether to use the real MLX embedding server
-            instead of mock embeddings. Defaults to False for backwards
-            compatibility and test determinism.
-        embedding_server_url: URL of the MLX embedding server. REQUIRED when
-            use_real_embeddings is True. Must be provided explicitly from
-            environment variable OMNIMEMORY__EMBEDDING__SERVER_URL.
-        embedding_timeout_seconds: Timeout for embedding requests in seconds.
-        embedding_max_retries: Maximum retries for embedding requests.
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    embedding_dimension: int = Field(
-        default=1024,
-        description="Dimension of embedding vectors",
-    )
-    simulate_latency_ms: int = Field(
-        default=0,
-        ge=0,
-        description="Simulated latency in milliseconds",
-    )
-    use_real_embeddings: bool = Field(
-        default=False,
-        description="Use real MLX embedding server instead of mock embeddings",
-    )
-    embedding_server_url: str | None = Field(
-        default=None,
-        description=(
-            "URL of MLX embedding server - REQUIRED when use_real_embeddings=True"
-        ),
-    )
-    embedding_timeout_seconds: float = Field(
-        default=5.0,
-        gt=0,
-        description="Timeout for embedding requests in seconds",
-    )
-    embedding_max_retries: int = Field(
-        default=2,
-        ge=0,
-        le=5,
-        description="Maximum retries for embedding requests",
-    )
 
 
 class HandlerQdrantMock:
