+++ src/omnimemory/nodes/memory_retrieval_effect/models/__init__.py
@@ -2,12 +2,18 @@
 # Copyright (c) 2025 OmniNode Team
 """Memory Retrieval Effect models.
 
-This package contains request/response models for the memory_retrieval_effect node.
+This package contains request/response models and configuration models
+for the memory_retrieval_effect node.
 
 .. versionadded:: 0.1.0
     Initial implementation for OMN-1387.
 """
 
+from .model_embedding_client_config import ModelEmbeddingClientConfig
+from .model_handler_db_mock_config import ModelHandlerDbMockConfig
+from .model_handler_graph_mock_config import ModelHandlerGraphMockConfig
+from .model_handler_memory_retrieval_config import ModelHandlerMemoryRetrievalConfig
+from .model_handler_qdrant_mock_config import ModelHandlerQdrantMockConfig
 from .model_memory_retrieval_request import ModelMemoryRetrievalRequest
 from .model_memory_retrieval_response import (
     ModelMemoryRetrievalResponse,
@@ -15,7 +21,14 @@ from .model_memory_retrieval_response import (
 )
 
 __all__ = [
+    # Request/Response models
     "ModelMemoryRetrievalRequest",
     "ModelMemoryRetrievalResponse",
     "ModelSearchResult",
+    # Configuration models
+    "ModelEmbeddingClientConfig",
+    "ModelHandlerDbMockConfig",
+    "ModelHandlerGraphMockConfig",
+    "ModelHandlerMemoryRetrievalConfig",
+    "ModelHandlerQdrantMockConfig",
 ]
