+++ src/omnimemory/nodes/memory_storage_effect/models/__init__.py
@@ -2,13 +2,17 @@
 # Copyright (c) 2025 OmniNode Team
 """Memory Storage Effect Models.
 
-Request and response models for memory storage CRUD operations.
+Request, response, and configuration models for memory storage CRUD operations.
 """
 
+from .model_filesystem_adapter_config import ModelFileSystemAdapterConfig
 from .model_memory_storage_request import ModelMemoryStorageRequest
 from .model_memory_storage_response import ModelMemoryStorageResponse
 
 __all__ = [
+    # Request/Response models
     "ModelMemoryStorageRequest",
     "ModelMemoryStorageResponse",
+    # Configuration models
+    "ModelFileSystemAdapterConfig",
 ]
