+++ src/omnimemory/nodes/similarity_compute/models/__init__.py
@@ -2,16 +2,21 @@
 # Copyright (c) 2025 OmniNode Team
 """Similarity Compute models.
 
-This package contains request/response models for the similarity_compute node.
+This package contains request/response models and configuration models
+for the similarity_compute node.
 
 .. versionadded:: 0.1.0
     Initial implementation for OMN-1388.
 """
 
+from .model_handler_similarity_compute_config import ModelHandlerSimilarityComputeConfig
 from .model_similarity_compute_request import ModelSimilarityComputeRequest
 from .model_similarity_compute_response import ModelSimilarityComputeResponse
 
 __all__ = [
+    # Request/Response models
     "ModelSimilarityComputeRequest",
     "ModelSimilarityComputeResponse",
+    # Configuration models
+    "ModelHandlerSimilarityComputeConfig",
 ]
