+++ src/omnimemory/utils/health_manager.py
@@ -15,7 +15,6 @@ import os
 import sys
 import time
 from datetime import datetime, timezone
-from enum import Enum
 from typing import TYPE_CHECKING, Literal, cast
 
 if TYPE_CHECKING:
@@ -32,13 +31,21 @@ except ImportError:
     pass
 
 import structlog
-from pydantic import BaseModel, ConfigDict, Field
 
 from ..models.foundation.model_health_metadata import HealthCheckMetadata
 from ..models.foundation.model_health_response import (
     ModelCircuitBreakerStats,
     ModelCircuitBreakerStatsCollection,
     ModelRateLimitedHealthCheckResponse,
+)
+from ..models.utils.model_health import (
+    DependencyType,
+    HealthCheckConfig,
+    HealthCheckDetails,
+    HealthCheckResult,
+    HealthStatus,
+    ResourceHealthCheck,
+    SystemHealth,
 )
 from .error_sanitizer import SanitizationLevel, sanitize_error
 
@@ -138,13 +145,12 @@ def _get_environment() -> str:
 
 
 from ..models.foundation.model_health_response import (
-    ModelDependencyStatus,
     ModelHealthResponse,
     ModelResourceMetrics,
 )
 from .concurrency import CircuitBreaker, CircuitBreakerOpenError
 from .observability import OperationType, correlation_context, trace_operation
-from .resource_manager import AsyncCircuitBreaker, CircuitBreakerConfig
+from .resource_manager import AsyncCircuitBreaker
 
 logger = structlog.get_logger(__name__)
 
@@ -169,72 +175,8 @@ __all__ = [
 ]
 
 
-class HealthStatus(Enum):
-    """Enhanced health status enumeration."""
-
-    HEALTHY = "healthy"
-    DEGRADED = "degraded"
-    UNHEALTHY = "unhealthy"
-    UNKNOWN = "unknown"
-    TIMEOUT = "timeout"
-    RATE_LIMITED = "rate_limited"
-    CIRCUIT_OPEN = "circuit_open"
-
-
-class DependencyType(Enum):
-    """Types of system dependencies."""
-
-    DATABASE = "database"
-    CACHE = "cache"
-    VECTOR_DB = "vector_db"
-    EXTERNAL_API = "external_api"
-    MESSAGE_QUEUE = "message_queue"
-    STORAGE = "storage"
-
-
-class HealthCheckConfig(BaseModel):
-    """Configuration for individual health checks."""
-
-    name: str = Field(description="Dependency name")
-    dependency_type: DependencyType = Field(description="Type of dependency")
-    timeout: float = Field(default=5.0, description="Health check timeout in seconds")
-    critical: bool = Field(
-        default=True, description="Whether failure affects overall health"
-    )
-    circuit_breaker_config: CircuitBreakerConfig | None = Field(default=None)
-    metadata: HealthCheckMetadata = Field(default_factory=HealthCheckMetadata)
-
-
-class HealthCheckResult(BaseModel):
-    """Result of an individual health check."""
-
-    config: HealthCheckConfig = Field(description="Health check configuration")
-    status: HealthStatus = Field(description="Health status")
-    latency_ms: float = Field(description="Check latency in milliseconds")
-    timestamp: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    error_message: str | None = Field(default=None)
-    metadata: HealthCheckMetadata = Field(default_factory=HealthCheckMetadata)
-
-    def to_dependency_status(self) -> ModelDependencyStatus:
-        """Convert to ModelDependencyStatus for API response."""
-        # Map HealthStatus to the expected Literal type
-        status_map: dict[HealthStatus, Literal["healthy", "degraded", "unhealthy"]] = {
-            HealthStatus.HEALTHY: "healthy",
-            HealthStatus.DEGRADED: "degraded",
-            HealthStatus.UNHEALTHY: "unhealthy",
-            HealthStatus.UNKNOWN: "unhealthy",
-            HealthStatus.TIMEOUT: "unhealthy",
-            HealthStatus.RATE_LIMITED: "degraded",
-            HealthStatus.CIRCUIT_OPEN: "degraded",
-        }
-        mapped_status = status_map.get(self.status, "unhealthy")
-        return ModelDependencyStatus(
-            name=self.config.name,
-            status=mapped_status,
-            latency_ms=self.latency_ms,
-            last_check=self.timestamp,
-            error_message=self.error_message,
-        )
+# HealthStatus, DependencyType, HealthCheckConfig, HealthCheckResult
+# are imported from ..models.utils.model_health
 
 
 class HealthCheckManager:
@@ -927,79 +869,8 @@ HealthCheckResultDict = dict[str, HealthCheckResultValue]
 
 
 # === TEST-COMPATIBLE INTERFACES ===
-# These classes provide the interface expected by test_health_manager.py
-
-
-class HealthCheckDetails(BaseModel):
-    """Strongly typed health check details with rate-limit and circuit tracking."""
-
-    message: str | None = Field(
-        default=None, description="Human-readable status message"
-    )
-    error: str | None = Field(default=None, description="Error message if unhealthy")
-    version: str | None = Field(default=None, description="Service version")
-    connection_url: str | None = Field(default=None, description="Connection URL")
-    last_check: str | None = Field(default=None, description="Last check timestamp")
-    latency_ms: float | None = Field(
-        default=None, description="Latency in milliseconds"
-    )
-    # Rate limiting state
-    rate_limit_active: bool = Field(
-        default=False, description="Whether rate limiting is currently active"
-    )
-    rate_limit_remaining: int | None = Field(
-        default=None, description="Remaining requests in current window"
-    )
-    rate_limit_reset_time: float | None = Field(
-        default=None, description="Time when rate limit resets (epoch)"
-    )
-    # Circuit breaker state
-    circuit_open: bool = Field(
-        default=False, description="Whether circuit breaker is open"
-    )
-    circuit_state: str | None = Field(
-        default=None, description="Current circuit breaker state"
-    )
-    circuit_failure_count: int | None = Field(
-        default=None, description="Number of failures recorded"
-    )
-    # Result details
-    result_type: str | None = Field(
-        default=None, description="Type of result (success/error/timeout)"
-    )
-    extra: dict[str, str] = Field(
-        default_factory=dict, description="Additional string details"
-    )
-
-
-class ResourceHealthCheck(BaseModel):
-    """Result of a resource health check."""
-
-    model_config = ConfigDict(use_enum_values=False)
-
-    status: HealthStatus = Field(description="Health status of the resource")
-    response_time: float = Field(default=0.0, description="Response time in seconds")
-    details: HealthCheckDetails = Field(
-        default_factory=HealthCheckDetails, description="Additional details"
-    )
-    correlation_id: str | None = Field(
-        default=None, description="Correlation ID for tracking"
-    )
-
-
-class SystemHealth(BaseModel):
-    """Overall system health status."""
-
-    model_config = ConfigDict(use_enum_values=False)
-
-    overall_status: HealthStatus = Field(description="Overall system health status")
-    resource_statuses: dict[str, ResourceHealthCheck] = Field(
-        default_factory=dict, description="Health status of individual resources"
-    )
-    timestamp: datetime = Field(
-        default_factory=lambda: datetime.now(timezone.utc),
-        description="Timestamp of health check",
-    )
+# HealthCheckDetails, ResourceHealthCheck, and SystemHealth
+# are imported from ..models.utils.model_health
 
 
 class HealthManager:
