+++ validate_foundation_isolated.py
@@ -48,7 +48,7 @@ def validate_protocol_definitions() -> dict[str, Any]:
         }
 
     except Exception as e:
-        print(f"❌ Protocol validation failed: {str(e)}")
+        print(f"❌ Protocol validation failed: {e!s}")
         traceback.print_exc()
         return {"success": False, "error": str(e)}
 
@@ -108,7 +108,7 @@ def validate_data_model_definitions() -> dict[str, Any]:
         }
 
     except Exception as e:
-        print(f"❌ Data model validation failed: {str(e)}")
+        print(f"❌ Data model validation failed: {e!s}")
         traceback.print_exc()
         return {"success": False, "error": str(e)}
 
@@ -144,7 +144,7 @@ def validate_error_model_definitions() -> dict[str, Any]:
         }
 
     except Exception as e:
-        print(f"❌ Error model validation failed: {str(e)}")
+        print(f"❌ Error model validation failed: {e!s}")
         traceback.print_exc()
         return {"success": False, "error": str(e)}
 
@@ -158,7 +158,7 @@ def validate_contract_specification() -> dict[str, Any]:
         if not contract_path.exists():
             return {"success": False, "error": "contract.yaml not found"}
 
-        with open(contract_path, "r") as f:
+        with open(contract_path) as f:
             contract = yaml.safe_load(f)
 
         # Validate contract structure
@@ -195,7 +195,7 @@ def validate_contract_specification() -> dict[str, Any]:
         }
 
     except Exception as e:
-        print(f"❌ Contract validation failed: {str(e)}")
+        print(f"❌ Contract validation failed: {e!s}")
         traceback.print_exc()
         return {"success": False, "error": str(e)}
 
@@ -242,7 +242,7 @@ def validate_project_structure() -> dict[str, Any]:
         }
 
     except Exception as e:
-        print(f"❌ Project structure validation failed: {str(e)}")
+        print(f"❌ Project structure validation failed: {e!s}")
         traceback.print_exc()
         return {"success": False, "error": str(e)}
